requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
anyio==4.13.0
attrs==26.1.0
autocommand==2.2.2
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.7
frozenlist==1.8.0
gql==4.0.0
graphql-core==3.2.8
idna==3.11
monarchmoneycommunity==1.3.1
multidict==6.7.1
oathtool==2.4.0
oauthlib==3.3.1
path==17.1.1
propcache==0.4.1
python-dotenv==1.2.2
requests==2.33.1
requests-oauthlib==1.3.1
splitwise==3.0.0
urllib3==2.6.3
yarl==1.23.0
